GtkWidget: Avoid widget->priv->alloc_needed when propagating exposes
It turns out that at various times we currently propagate draws
when the child has alloc_needed set. We need to fix these, but
we don't want that to block the current work, so for now we just
avoid that assert when we propagate draws (its what we used to
do after all).