GtkTreeView: Invalidate pixel cache on queue_draw
Whenever queue_draw is called on the treeview we invalidate the
whole pixel cache, as there is no way we can know when the area
outside the visible region should be updated from the region
(as it is in widget coords).

This is not typically a problem though, we only queue redraws
on the treeview itself on e.g. resizes or restyles, which are
generally rare.