Update Visual Studio 2010 projects
Add a PlatformToolset tag for each configuration for project files that
do not yet have them.  This is to ease support for Visual Studio 2012 as
we can copy and easily replace a few items with automated scripts as
project files for Visual Studio 2010 and 2012 are very similar.

This might change when we eventually support the Metro (aka Windows 8
Modern UI), but this will suffice for the time being.