GtkFileChooser: Missing hour on today's files for it&hu locales on Win32
The file chooser does not show the hours for files modified today
when using it & hu locales on Win32.  This is due to the %k format conversion
for strftime() that is in Olson's Timezone Package, which msvcrt.dll doesn't have.
Cygwin has it, though.

So for now, we just use a %H format.  While %k would render
six-in-the-morning as " 6" (6 preceded by a space), %H will render it
as "06".

https://bugzilla.gnome.org/show_bug.cgi?id=560337